Remove unnecessary and inaccurate extension checks
We can't accurately judge a file's content based solely on its extension. To do that, we would need to parse it. But hey, we already have logic in place that will throw an exception if the provided file isn't valid JSON or TOML, as expected. And that exception will be far more descriptive than our generic "wrong file extension!" error.